+++ graphql-dynamodb-manager-core/src/main/java/com/fleetpin/graphql/dynamodb/manager/access/ForbiddenWriteException.java
@@ -10,16 +10,10 @@
  * the License.
  */
 
-package com.fleetpin.graphql.dynamodb.manager;
+package com.fleetpin.graphql.dynamodb.manager.access;
 
-import static java.lang.annotation.RetentionPolicy.RUNTIME;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.Target;
-
-@Retention(RUNTIME)
-@Target(ElementType.TYPE)
-public @interface TableName {
-	public String value();
+public class ForbiddenWriteException extends RuntimeException {
+	public ForbiddenWriteException(String msg) {
+		super(msg);
+	}
 }
