+++ graphql-dynamodb-manager-memory/src/test/java/com/fleetpin/graphql/dynamodb/manager/memory/InMemoryDynamoDbTest.java
@@ -21,7 +21,11 @@ import com.fasterxml.jackson.datatype.jdk8.Jdk8Module;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import com.fasterxml.jackson.module.paramnames.ParameterNamesModule;
 import com.fleetpin.graphql.builder.annotations.Id;
-import com.fleetpin.graphql.dynamodb.manager.*;
+import com.fleetpin.graphql.dynamodb.manager.annotations.GlobalIndex;
+import com.fleetpin.graphql.dynamodb.manager.annotations.SecondaryIndex;
+import com.fleetpin.graphql.dynamodb.manager.DatabaseKey;
+import com.fleetpin.graphql.dynamodb.manager.DatabaseQueryKey;
+import com.fleetpin.graphql.dynamodb.manager.table.Table;
 import com.google.common.collect.HashMultimap;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
