+++ graphql-dynamodb-manager-test/src/test/java/com/fleetpin/graphql/dynamodb/manager/test/DynamoDbIndexesTest.java
@@ -12,15 +12,16 @@
 
 package com.fleetpin.graphql.dynamodb.manager.test;
 
-import com.fleetpin.graphql.dynamodb.manager.GlobalIndex;
-import com.fleetpin.graphql.dynamodb.manager.SecondaryIndex;
-import com.fleetpin.graphql.dynamodb.manager.Table;
+import com.fleetpin.graphql.dynamodb.manager.annotations.GlobalIndex;
+import com.fleetpin.graphql.dynamodb.manager.annotations.SecondaryIndex;
+import com.fleetpin.graphql.dynamodb.manager.table.Table;
 import com.fleetpin.graphql.dynamodb.manager.dynamo.Database;
+import com.fleetpin.graphql.dynamodb.manager.test.annotations.TestDatabase;
 import org.junit.jupiter.api.Assertions;
 
 import java.util.concurrent.ExecutionException;
 
-final class DynamoDBIndexesTest {
+final class DynamoDbIndexesTest {
 
 	@TestDatabase
 	void testGlobal(final Database db) throws InterruptedException, ExecutionException {
