+++ graphql-dynamodb-manager-test/src/test/java/com/fleetpin/graphql/dynamodb/manager/test/DynamoDbPutGetDeleteTest.java
@@ -12,13 +12,14 @@
 
 package com.fleetpin.graphql.dynamodb.manager.test;
 
-import com.fleetpin.graphql.dynamodb.manager.Table;
+import com.fleetpin.graphql.dynamodb.manager.table.Table;
 import com.fleetpin.graphql.dynamodb.manager.dynamo.Database;
+import com.fleetpin.graphql.dynamodb.manager.test.annotations.TestDatabase;
 import org.junit.jupiter.api.Assertions;
 
 import java.util.concurrent.ExecutionException;
 
-final class DynamoDBPutGetDeleteTest {
+final class DynamoDbPutGetDeleteTest {
 
 	@TestDatabase
 	void testSimplePutGetDelete(final Database db) throws InterruptedException, ExecutionException {
