+++ graphql-dynamodb-manager-test/src/test/java/com/fleetpin/graphql/dynamodb/manager/test/DynamoDbQueryTest.java
@@ -12,14 +12,15 @@
 
 package com.fleetpin.graphql.dynamodb.manager.test;
 
-import com.fleetpin.graphql.dynamodb.manager.Table;
+import com.fleetpin.graphql.dynamodb.manager.table.Table;
 import com.fleetpin.graphql.dynamodb.manager.dynamo.Database;
+import com.fleetpin.graphql.dynamodb.manager.test.annotations.TestDatabase;
 import org.junit.jupiter.api.Assertions;
 
 import java.util.Comparator;
 import java.util.concurrent.ExecutionException;
 
-final class DynamoDBQueryTest {
+final class DynamoDbQueryTest {
 	@TestDatabase
 	void testSimpleQuery(final Database db) throws InterruptedException, ExecutionException {
 		db.put(new SimpleTable("garry")).get();
